Moving to web/ (#15)
Moving web-view assets to web/. Currently that's only preferences but
will eventually include sidebar and overlay.

Main motivation for this is because of different APIs available between
web views and JavaScriptCore used in the entry.cjs file. `URL()` is not
available in entry.cjs but is in web view cjs files. Will eventually
have a lib-web.cjs for shared code among web views.

Keeping use of cjs instead of js to silence VS Code suggestions.