postMessageAck and onMessageAck (#20)
There is a race condition when entry tries to post messages to the
sidebar. Sometimes the message is sent before the sidebar is done
initializing. Mitigating this with wrapper functions that acknowledge
messages.

Re-defining onFileLoaded to just send messages to the sidebar. Will
eventually do the same with an overlay.